+++ src/models/temporal/memory/controllers/gddr6_controller.cpp
@@ -338,7 +338,7 @@ void GDDR6MemoryController::decode_address(
     uint64_t addr = address >> byte_offset_bits;
 
     if (channel_bits > 0) {
-        channel = addr & ((1 << channel_bits) - 1);
+        channel = static_cast<uint8_t>(addr & ((1 << channel_bits) - 1));
         addr >>= channel_bits;
     } else {
         channel = 0;
@@ -347,7 +347,7 @@ void GDDR6MemoryController::decode_address(
     col = addr & ((1 << col_bits) - 1);
     addr >>= col_bits;
 
-    bank = addr & ((1 << bank_bits) - 1);
+    bank = static_cast<uint8_t>(addr & ((1 << bank_bits) - 1));
     addr >>= bank_bits;
 
     row = addr & ((1 << gddr6_config_.row_bits) - 1);
